Revert "Unify server-side and model-side Config(Part-5) (#3497)"
This reverts commit c389a4013cbf596c0461b2c52a7a8a19bf2dd131.
