--- tests/utils/test_config.py
@@ -1,81 +1,0 @@
-import unittest
-
-from fastdeploy.config import (
-    CacheConfig,
-    FDConfig,
-    GraphOptimizationConfig,
-    ParallelConfig,
-)
-
-
-class TestConfig(unittest.TestCase):
-    def test_fdconfig_nnode(self):
-        parallel_config = ParallelConfig({"tensor_parallel_size": 16, "expert_parallel_size": 1})
-        graph_opt_config = GraphOptimizationConfig({})
-        cache_config = CacheConfig({})
-        fd_config = FDConfig(
-            parallel_config=parallel_config,
-            graph_opt_config=graph_opt_config,
-            cache_config=cache_config,
-            ips=["1.1.1.1", "0.0.0.0"],
-            test_mode=True,
-        )
-        assert fd_config.nnode == 2
-        assert fd_config.is_master is False
-
-    def test_fdconfig_ips(self):
-        parallel_config = ParallelConfig({})
-        graph_opt_config = GraphOptimizationConfig({})
-        cache_config = CacheConfig({})
-        fd_config = FDConfig(
-            parallel_config=parallel_config,
-            graph_opt_config=graph_opt_config,
-            cache_config=cache_config,
-            ips="0.0.0.0",
-            test_mode=True,
-        )
-        assert fd_config.master_ip == "0.0.0.0"
-
-    def test_fdconfig_max_num_tokens(self):
-        parallel_config = ParallelConfig({})
-        graph_opt_config = GraphOptimizationConfig({})
-        cache_config = CacheConfig({})
-        cache_config.enable_chunked_prefill = True
-        fd_config = FDConfig(
-            parallel_config=parallel_config,
-            graph_opt_config=graph_opt_config,
-            cache_config=cache_config,
-            ips="0.0.0.0",
-            test_mode=True,
-        )
-        assert fd_config.max_num_batched_tokens == 2048
-
-        cache_config.enable_chunked_prefill = False
-        fd_config = FDConfig(
-            parallel_config=parallel_config,
-            graph_opt_config=graph_opt_config,
-            cache_config=cache_config,
-            ips="0.0.0.0",
-            test_mode=True,
-        )
-        assert fd_config.max_num_batched_tokens == 8192
-
-    def test_fdconfig_init_cache(self):
-        parallel_config = ParallelConfig({})
-        graph_opt_config = GraphOptimizationConfig({})
-        cache_config = CacheConfig({})
-        cache_config.cache_transfer_protocol = "rdma,ipc"
-        cache_config.pd_comm_port = "2334"
-        fd_config = FDConfig(
-            parallel_config=parallel_config,
-            graph_opt_config=graph_opt_config,
-            cache_config=cache_config,
-            splitwise_role="prefill",
-            test_mode=True,
-        )
-        fd_config.init_cache_info()
-        assert fd_config.disaggregate_info["role"] == "prefill"
-
-
-if __name__ == "__main__":
-    unittest.main()
